Extract parsers into own files
