+++ lib/openapi_contracts/parser/transformers.rb
@@ -1,0 +1,5 @@
+module OpenapiContracts::Parser::Transformers
+  autoload :Base,     'openapi_contracts/parser/transformers/base'
+  autoload :Nullable, 'openapi_contracts/parser/transformers/nullable'
+  autoload :Pointer,  'openapi_contracts/parser/transformers/pointer'
+end
